+++ Changelog.md
@@ -1,6 +1,12 @@
+# Release 0.1.0
+
+## Features
+
+- add clear, get_return_value, get_all_tasks method to DAG class
+
 # Release 0.0.1
 
 ## Features
 
 - First release
-  - add operator and module for function scheduling
+    - add operator and module for function scheduling